pyproject.toml
[project]
name = "lang"
version = "0.1.0"
description = ""
authors = [
    {name = "Siddhant Gupta"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "langchain (>=0.3.27,<0.4.0)",
    "openai (>=1.98.0,<2.0.0)",
    "langchain-openai (>=0.3.28,<0.4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
